backend/requirements.txt
# Python Dependencies for Backend

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
# psycopg2-binary==2.9.9  # Commented out - using SQLite for development
alembic==1.12.1

# Async & Task Queue
httpx==0.25.1
aiofiles==23.2.1

# AI/ML Libraries
openai==1.3.5
anthropic==0.7.1
google-generativeai==0.3.1

# NLP & Text Processing
spacy==3.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0

# Data Processing
pandas==2.1.3
openpyxl==3.1.2
numpy==1.26.2
scikit-learn==1.3.2

# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Redis (for caching)
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
